Refactor packages incorrect paths
